refactor: ✨ rename PageHome to PageDetails
