+++ src/ui/pages/PageDetails.vue
