+++ src/ui/router/routes.ts
@@ -1,15 +1,15 @@
 import type { RouteRecordRaw } from 'vue-router'
 
-import PageHome from '@/pages/PageHome.vue'
+import PageDetails from '@/pages/PageDetails.vue'
 
 export const routes: RouteRecordRaw[] = [
   {
-    path: '/',
-    name: 'PageHome',
-    component: PageHome
+    path: '/details',
+    name: 'PageDetails',
+    component: PageDetails
   },
   {
     path: '/:pathMatch(.*)*',
-    redirect: { name: 'PageHome' }
+    redirect: { name: 'PageDetails' }
   }
 ]
